Added a JavaFX example (#35)
At Fosdem I'll give a talk in the "Tool the Docs" room about a
chat-search-implementation based on a docs set. For this demo I created
a JavaFX application and would like to contribute the basic UI of that
demo to the Langchain4J project.

See the included screenshot for what this application provides.