WIP: agentic tutorial (#168)
Hi,

I'd like to have a dedicated module in the examples for a
langchain4j-agentic tutorial to
- make an easy-to-follow step-by-step tutorial that developers can run
- cover different options (untyped/typed agents, output parsing,
outputting parts of the AgentScope, ...)
- show intricacies beyond what is explained in the documentation, where
I think developers will wonder (eg. saving intermediary variable states,
re-injecting POJOs in the next step, ...)

Here's a proposal for the first 3 classes (basic agent, sequence, loop)
that I plan to extend in the coming days.
I tried to give a realistic useful example that I elaborate on and that
allows to demo from very basic to a complex compound system: a
CvGenerator (from a life story), a CvTailor (tailors to a job
description), a CvReviewer that gives feedback, etc.

@dliubarskyi do you already want to have a look or wait till I'm done?